requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
python-docx==1.0.1
PyPDF2==3.0.1
markdown==3.5.1
pydantic==2.4.2
sqlalchemy==2.0.23
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0.post1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 